refactor: Move numbers nodes to individual files
Implements ML-597
